Squashed commit of the following:
commit 5cbbdcedca625852772a7e31fa0092b8466fb0ab
Author: Line Colin <line.colin@ens-lyon.fr>
Date:   Fri May 16 17:05:14 2025 +0200

    Add the option for uniform or radius dependant gravity acceleration

This update introduces the option of choosing between uniform and radius-dependent gravity acceleration in the SphStability class.

The gravity acceleration that depends of the radius is defined as:

	g(r) = - g_0 * r * e_r

The 'uniform_gravity' boolean flag has been added to the dataclass to control this option. When set to 'False', gravity acceleration varies linearly with radius.

This change required modifications in 'lmat'.